adds ranking leaderboard
